convert prisma to lib, expose internals for using prisma as rust lib
