+++ query-engine/prisma/src/main.rs
@@ -3,44 +3,17 @@ extern crate log;
 #[macro_use]
 extern crate rust_embed;
 
-use cli::*;
-use error::*;
-use once_cell::sync::Lazy;
-use opt::*;
-use request_handlers::{PrismaRequest, PrismaResponse, RequestHandler};
-use server::{HttpServer, HttpServerBuilder};
-use std::{convert::TryFrom, error::Error, net::SocketAddr, process};
+use prisma::{
+    AnyError, set_panic_hook, init_logger, server::{HttpServerBuilder, HttpServer},
+    error::PrismaError, cli::CliCommand, PrismaResult, opt::PrismaOpt,
+};
+use std::process;
 use structopt::StructOpt;
-use tracing::subscriber;
-use tracing_log::LogTracer;
-use tracing_subscriber::{EnvFilter, FmtSubscriber};
-
-mod cli;
-mod context;
-mod dmmf;
-mod error;
-mod exec_loader;
-mod opt;
-mod request_handlers;
-mod server;
+use std::net::SocketAddr;
+use std::convert::TryFrom;
 
 #[cfg(test)]
 mod tests;
-
-#[derive(Debug, Clone, PartialEq, Copy)]
-pub enum LogFormat {
-    Text,
-    Json,
-}
-
-static LOG_FORMAT: Lazy<LogFormat> =
-    Lazy::new(|| match std::env::var("RUST_LOG_FORMAT").as_ref().map(|s| s.as_str()) {
-        Ok("devel") => LogFormat::Text,
-        _ => LogFormat::Json,
-    });
-
-pub type PrismaResult<T> = Result<T, PrismaError>;
-type AnyError = Box<dyn Error + Send + Sync + 'static>;
 
 #[tokio::main]
 async fn main() -> Result<(), AnyError> {
@@ -95,65 +68,6 @@ async fn main() -> Result<(), AnyError> {
             info!("Encountered error during initialization:");
             err.render_as_json().expect("error rendering");
             process::exit(1);
-        }
-    }
-
-    Ok(())
-}
-
-fn init_logger() -> Result<(), AnyError> {
-    LogTracer::init()?;
-
-    match *LOG_FORMAT {
-        LogFormat::Text => {
-            let subscriber = FmtSubscriber::builder()
-                .with_env_filter(EnvFilter::from_default_env())
-                .finish();
-
-            subscriber::set_global_default(subscriber)?;
-        }
-        LogFormat::Json => {
-            let subscriber = FmtSubscriber::builder()
-                .json()
-                .with_env_filter(EnvFilter::from_default_env())
-                .finish();
-
-            subscriber::set_global_default(subscriber)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn set_panic_hook() -> Result<(), AnyError> {
-    match *LOG_FORMAT {
-        LogFormat::Text => (),
-        LogFormat::Json => {
-            std::panic::set_hook(Box::new(|info| {
-                let payload = info
-                    .payload()
-                    .downcast_ref::<String>()
-                    .map(Clone::clone)
-                    .unwrap_or_else(|| info.payload().downcast_ref::<&str>().unwrap().to_string());
-
-                match info.location() {
-                    Some(location) => {
-                        tracing::event!(
-                            tracing::Level::ERROR,
-                            message = "PANIC",
-                            reason = payload.as_str(),
-                            file = location.file(),
-                            line = location.line(),
-                            column = location.column(),
-                        );
-                    }
-                    None => {
-                        tracing::event!(tracing::Level::ERROR, message = "PANIC", reason = payload.as_str());
-                    }
-                }
-
-                std::process::exit(255);
-            }));
         }
     }
